Flushing out the agent's view frustum.
