+++ agents_playground/core/types.py
@@ -24,22 +24,6 @@ Sample = int | float
 CanvasLocation = List[int] | Tuple[int, ...]
 CellLocation = Tuple[int, int]
 
-#Convenience tuples for working with grid coordinates. 
-CoordinateComponent = int | float
-
-class Coordinate(NamedTuple):
-  x: CoordinateComponent
-  y: CoordinateComponent
-
-  def multiply(self, p: Coordinate) -> Coordinate:
-    return Coordinate(self.x * p.x, self.y * p.y) 
-
-  def shift(self, p:Coordinate) -> Coordinate:
-    return Coordinate(self.x + p.x, self.y + p.y) 
-
-  def to_tuple(self) -> Tuple[CoordinateComponent, CoordinateComponent]:
-    return (self.x, self.y)
-
 # Handling Size
 Dimension = int | float
 
@@ -54,45 +38,3 @@ class Size:
 
   def scale(self, amount: float) -> Size:
     return Size(self.width * amount, self.height * amount)
-
-# Dealing with rotations
-Radians = float
-Degrees = float
-
-# Geometry
-# TODO: Should probably create a new folder that contains both this and Vector.
-class AABBox:
-  """An axis-aligned bounding box."""
-  def __init__(self, min: Coordinate, max: Coordinate) -> None:
-    """ Creates an axis-aligned bounding box defined by two points.
-    Args:
-      - min: The left most, upper point.
-      - max: The right most, lower point.
-    """
-    self._min: Coordinate = min
-    self._max: Coordinate = max
-
-  @property
-  def min(self) -> Coordinate:
-    return self._min
-  
-  @property
-  def max(self) -> Coordinate:
-    return self._max
-
-  def point_in(self, point: Coordinate) -> bool:
-    """Calculates if a given point is in the box."""
-    return (self._min.x <= point.x and point.x <= self._max.x) \
-      and (self._min.y <= point.y and point.y <= self._max.y)
-
-  def __repr__(self) -> str:
-    return f"""
-    {self.__class__.__name__}
-    min: {self._min}
-    max: {self._max}
-    """
-  
-class EmptyAABBox(AABBox):
-  """Convince class for creating a bounding box with no size."""
-  def __init__(self) -> None:
-    super().__init__(Coordinate(0,0), Coordinate(0,0))
